+++ homecare_frontend/src/pages/AdminUserCreate/AdminUserCreateForm.js
@@ -4,7 +4,7 @@ import { Button, FormHelperText, Grid, InputLabel, MenuItem, OutlinedInput, Stac
 // third party
 import * as Yup from 'yup';
 import { Field, Formik } from 'formik';
-import FormikCustomSelect from 'components/form/FormikCustomSelect';
+import FormikCustomSelect from 'components/customFormFields/FormikCustomSelectField';
 import { professions } from 'constants/professions';
 import PropTypes from 'prop-types';
